MoreNewBookIndex页面排序功能完成但是分页功能还待开发

+++ src/main/java/com/bai/pojo/bo/BookQueryBo.java
@@ -30,7 +30,7 @@ public class BookQueryBo {
     private Integer alllocflag;
     private Character classno;
     private Long multiclassno;
-    private Long callnorderid;
+    private Integer callnorderid;
     private Integer orderid;
     private String searchtxt;
     private Integer searchfield;
